+++ CHANGELOG.md
@@ -17,7 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - gateway: Check RacksDB version executed by agent is greater or equal to the
   minimal supported version specified in gateway configuration settings
   (#415→#417).
-- agent: Skip registering of RacksDB API endpoints when disabled (#440).
+- agent:
+  - Skip registering of RacksDB API endpoints when disabled (#440).
+  - Report error in logs instead of failing with critical error when unable to
+    load RacksDB database (#458).
 - frontend:
   - Reduce height of error message container when unable to retrieve
     infrastructure graphical representation from RacksDB in resources page.